eva-fees-receipts/requirements.txt
mysql-connector-python
python-docx
num2words
docx2pdf